+++ client/coral-embed-stream/src/actions/commentBox.js
