+++ client/coral-embed-stream/src/constants/commentBox.js
@@ -1,0 +1,5 @@
+const prefix = 'TALK_EMBED_STREAM_COMMENT_BOX';
+
+export const ADD_TAG = `${prefix}_ADD_TAG`;
+export const REMOVE_TAG = `${prefix}_REMOVE_TAG`;
+export const CLEAR_TAGS = `${prefix}_CLEAR_TAGS`;
