+++ client/coral-embed-stream/src/tabs/stream/components/InfoBox.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Markdown from 'coral-framework/components/Markdown';
 
+// TODO: remove this.
 const packagename = 'talk-plugin-infobox';
 
 const InfoBox = ({ enable, content }) => (
