+++ client/coral-embed-stream/src/tabs/stream/components/TagLabel.js
@@ -1,5 +1,7 @@
 import React from 'react';
 
+// TODO: change className.
+
 const TagLabel = ({ children }) => (
   <div className="talk-plugin-tag-label">{children}</div>
 );
