--- client/talk-plugin-commentbox/constants.js
@@ -1,3 +1,0 @@
-export const ADD_TAG = 'ADD_TAG';
-export const REMOVE_TAG = 'REMOVE_TAG';
-export const CLEAR_TAGS = 'CLEAR_TAGS';
