--- client/talk-plugin-commentbox/index.js
@@ -1,1 +1,0 @@
-export { default as reducer } from './reducer';
