+++ client/talk-plugin-replies/index.js
@@ -1,4 +1,1 @@
-import ReplyBox from './ReplyBox';
-import ReplyButton from './ReplyButton';
-
-export { ReplyBox, ReplyButton };
+export { default as ReplyButton } from './ReplyButton';
